+++ src/test/java/ru/netology/stats/service/MonthlySalesServiceTest.java
@@ -17,8 +17,8 @@ public class MonthlySalesServiceTest {
     public void TestAverageAmount() {
         MonthlySalesService service = new MonthlySalesService();
         long[] sales = {440_000, 352_000, 856_120, 297_340, 462_310, 58_700, 332_890, 764_100, 963_150, 852_340, 963_000, 498_000};
-        int expected = 569_995;
-        int actual = service.averageAmount(sales);
+        long expected = 569_995;
+        long actual = service.averageAmount(sales);
         Assertions.assertEquals(expected, actual);
 
     }
@@ -45,8 +45,8 @@ public class MonthlySalesServiceTest {
     public void TestLowSales() {
         MonthlySalesService service = new MonthlySalesService();
         long[] sales = {440_000, 352_000, 856_120, 297_340, 462_310, 58_700, 332_890, 764_100, 963_150, 852_340, 963_000, 498_000};
-        int expected = 7;
-        int actual = service.calcLowMonth(sales);
+        long expected = 7;
+        long actual = service.calcLowMonth(sales);
         Assertions.assertEquals(expected, actual);
     }
 
@@ -54,8 +54,8 @@ public class MonthlySalesServiceTest {
     public void TestHighSales() {
         MonthlySalesService service = new MonthlySalesService();
         long[] sales = {440_000, 352_000, 856_120, 297_340, 462_310, 58_700, 332_890, 764_100, 963_150, 852_340, 963_000, 498_000};
-        int expected = 5;
-        int actual = service.calcHighMonth(sales);
+        long expected = 5;
+        long actual = service.calcHighMonth(sales);
         Assertions.assertEquals(expected, actual);
     }
